New cache system
The current cache system is nit heavily used and inefficient in places. In addition, the code is IMHO somewhat
convoluted and hard to follow in places. So set a up a branch to see if we can simplify things a bit. This
introduces a new C++ based cache system, based on a simple hash map. lLooking up items in the cache should be
much faster now. A disadvantage of the new system may be that pointers into the cache are no longer stable, so
all access into the cache must be done by name. For code clarity, not having access to the internals of the cache
is a win however.