Make free_item function obligatory again.
There are currently 3 caches in use: the system cache containing other caches, the 3d object cache and the
texture cache. None of them free items from the cache in the course of maintenance. So, let's remove the ejection
of unused items from the cache from the maintenance function. Then we can make the free_item function mandatory
again, and always set it, so that item data is automatically freed when the cache is deleted. That way we don't
have to play games with setting a free handler right before the cache. This also removes the need to have a
function that sets a free handler.

If we ever want to use a cache again that does eject unused items occasionally (instead of just compacting them),
we should probably have a separate parameter (apart from _time_limit) that indicates how often this function
should be called. That way, we can still turn off item ejection for caches that don't need it.