Remove e3d object freeing from do_load_e3d_detail.
Currently the function do_load_e3d_detail(), which does the grunt work of loading e3d data, free()s the e3d
object it is working on in every error path. Setting aside for the moment the wisdom of freeing a structure that was
passed in as an argument, this is very error prone. So as a first step to cleaning this up, do not free the
pointer in do_load_e3d_detail(), but instead in its only caller load_e3d_detail(). Instead return the total
amount of memory allocated for the loaded data structures, or -1 for error. This way, we have a single central
place where the freeing is done.