Do not free e3d data in loading function.
Currently, the e3d_load_detail function will free the e3d data structure that it's trying to fill when an error
happens. This is problematic, because the data is already cached, and by freeing the structure the pointer in the
cache is invalidated. Worse, when called from load_e3d_detail_if_needed(), the result of loadin the data is never
checked, and so there is even no way to handle an error.

Let the caller be responsible for freeing e3d data if loading the details fails. Furthermore, propagate the
result in load_e3d_detail_if_needed() so that the caller knows when loading fails, and can take appropriate
action.