Initial version of new encyclopedia code
Rewrite current encyclopedia code in hopefully more understandable C++. The goal is to extend the
rewritten code to support a new, semantic, XML format in addition to the current legacy encyclopedia.
The current commit adds a new encyclopedia window. Not all functionality is included yet, only
display of the initial page.