Introduce C++ Window class
Create a new Window class that provides a C++ abstraction for client window information and that can install
event handlers automatically. Use it in the new encyclopedia window to get rid of a bunch of hand-rolled
static event handlers. As a side effect, the encyclopedia window no longer needs to be singleton.